Cleanup + finishing touches before deadline
Removed unnecessary comments.
Removed most of the code that wasn't being used.
Some parts such as crudOrder.js are left in for future implementation.
Moved functions that were spread out in files to a single file in a reasonable way, such as create, read etc into a crud file, and moved endpoints from app.js to routers to create a better strucutre.
Added a checkout functionality that removed beers from the database, but it still doesn't charge any money.
Added the company logo to the header and footer.